requirements.txt
dm_haiku==0.0.10
einops==0.7.0
flax==0.7.5
h5py==3.9.0
hydra-core==1.3.2
ipython==8.12.3
jax==0.4.23
matplotlib==3.7.3
numpy==1.24.4
omegaconf==2.3.0
optax==0.2.1
pandas==2.2.1
scipy==1.12.0
setuptools==65.5.0
tqdm==4.66.1
hydra-colorlog==1.2.0
hydra-submitit-launcher==1.2.0